+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 ## 0.5.1 (unreleased)
 
+  - Fixed `halt`, `destroy`, `reload` to where they failed if the VM was
+    in a saved state. [GH-123]
   - Added `config.chef.recipe_url` which allows you to specify a URL to
     a gzipped tar file for chef solo to download cookbooks. See the
     [chef-solo docs](http://wiki.opscode.com/display/chef/Chef+Solo#ChefSolo-RunningfromaURL) for more information.
+    [GH-121]
   - Added `vagrant box repackage` which repackages boxes which have
     been added. This is useful in case you want to redistribute a base
-    box you have but may have lost the actual "box" file.
+    box you have but may have lost the actual "box" file. [GH-120]
 
 ## Previous